+++ enclave/strings.c
@@ -11,17 +11,17 @@ size_t oel_strlen(const char* s)
         p += 6;
 
     if (!p[0])
-        return p - s;
+        return (size_t)(p - s);
     if (!p[1])
-        return p - s + 1;
+        return (size_t)(p - s + 1);
     if (!p[2])
-        return p - s + 2;
+        return (size_t)(p - s + 2);
     if (!p[3])
-        return p - s + 3;
+        return (size_t)(p - s + 3);
     if (!p[4])
-        return p - s + 4;
+        return (size_t)(p - s + 4);
     if (!p[5])
-        return p - s + 5;
+        return (size_t)(p - s + 5);
 
     /* Unreachable */
     return 0;
@@ -34,7 +34,7 @@ size_t oel_strnlen(const char* s, size_t n)
     while (n-- && *p)
         p++;
 
-    return p - s;
+    return (size_t)(p - s);
 }
 
 int oel_strcmp(const char* s1, const char* s2)
@@ -83,7 +83,7 @@ size_t oel_strlcpy(char* dest, const char* src, size_t size)
     while (*src)
         src++;
 
-    return src - start;
+    return (size_t)(src - start);
 }
 
 size_t oel_strlcat(char* dest, const char* src, size_t size)
@@ -194,7 +194,7 @@ void* oel_memset(void* s, int c, size_t n)
     }
 
     while (n--)
-        *p++ = c;
+        *p++ = (uint8_t)c;
 
     return s;
 }
